#567 Let ./bin/cifsynthanaly application wait ControlEditor to close.
To enable the standalone application ./bin/cifsynthanaly to wait for the
editor to complete the method public waitUntilClosed() has been added to
the ControlEditor. This method blocks the calling thread using wait().
The release the waiting thread a DisposeListener has been added which
will notify all waiting threads when the Editors content is disposed.