#611 Use format string with on-the-fly new method creation.
The CifFormatPatternCodeGenerator became trickier, as we now have layered fmt functions. 
Temporarily disabled non-static merge, because when we use a format string as input, we need the order of results to add. 'others' has an order, but unclear where 'this' to insert.
Currently outsourcing format string checking to the fmt method.